+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 * Ensured metrics are reported even for failed requests
 * Added `TryGetItemAsync` (same as `GetItemAsync`, but returns `None`, instead of throwing, if an item is not present)
 * Switched test framework to Xunit, assertions to Unquote, runner to `dotnet test` 
+* Added `TableContext.CreateUnverified` (`TableContext.CreateAsync` without the optional store round-trips) 
 
 ### 0.9.3-beta
 * Added `RequestMetrics` record type